newData

+++ app/src/main/java/com/soha/weather_app/weather/db/model/Current.kt
@@ -9,7 +9,7 @@ data class Current(
     @SerializedName("dew_point")
     val dewPoint: Double,
     @SerializedName("dt")
-    val dt: Double,
+    val dt: Int,
     @SerializedName("feels_like")
     val feelsLike: Double,
     @SerializedName("humidity")
@@ -25,7 +25,7 @@ data class Current(
     @SerializedName("uvi")
     val uvi: Double,
     @SerializedName("visibility")
-    val visibility: Double,
+    val visibility: Int,
     @SerializedName("weather")
     val weather: List<Weather>,
     @SerializedName("wind_deg")
